+++ src/Phan/Library/None.php
@@ -8,7 +8,7 @@ use Exception;
  * This represents the absence of a value in an Option.
  *
  * @inherits Option<null>
- * @immutable
+ * @phan-pure
  */
 class None extends Option
 {