+++ src/Phan/Library/Some.php
@@ -11,7 +11,7 @@ namespace Phan\Library;
  *
  * @inherits Option<T>
  * phpcs:disable PSR2.Classes.PropertyDeclaration.Underscore
- * @immutable
+ * @phan-pure
  */
 class Some extends Option
 {